+++ sbt/src/sbt-test/dependency-management/cache-update/build.sbt
@@ -22,7 +22,7 @@ TaskKey[Unit]("check") := {
   type In = IvyConfiguration :+: ModuleSettings :+: UpdateConfiguration :+: HNil
 
   import sbt.util.CacheImplicits._
-  import sbt.Classpaths.AltLibraryManagementCodec._
+  import sbt.internal.librarymanagement.AltLibraryManagementCodec._
 
   val f: In => Unit =
     Tracked.inputChanged(cacheStoreFactory make "inputs") { (inChanged: Boolean, in: In) =>
